pyproject.toml
[project]
name = "tron-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "base58>=2.1.1",
    "dependency-injector>=4.46.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "pydantic-settings>=2.8.1",
    "pytest-asyncio>=0.26.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.11.5",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.1",
]

[tool.ruff]
line-length = 100  
target-version = "py311"
exclude = ["src/alembic/"]

[tool.ruff.lint]
select = ["ALL"]  
ignore = [
    "E501", "D", "UP035", "ANN401",
    "B010", "B009", "SIM117", "EM101",
    "TRY003", "FAST002", "B008", "TID252",
    "PGH003"
]

[tool.mypy]
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = ["src", "tests"]
testpaths = ["tests"]
log_cli = true
log_cli_level = "INFO"

